more layout + theme optimization/refactor
