+++ src/components/AvatarArea/index.jsx
@@ -19,7 +19,6 @@ function AvatarArea({ user }) {
   return (
     <div className="avatarArea">
       <div className="avatarArea__clicker" onClick={handleClick}>
-        {/* NICE TRICK : again-> optional chaining using `?` -> if object is still null when trying to access its properties, at least there won't be a JS error */}
         <Avatar src={user?.images[0]?.url} alt={user?.display_name} />
         <h4>{user?.display_name}</h4>
       </div>
