+++ src/components/Login/Login.css
@@ -1,6 +1,6 @@
 .login {
   display: grid;
-  place-items: center; /* NICE TRICK */
+  place-items: center;
   height: 100vh;
   background-color: black;
 }
