+++ src/components/SongRow/SongRow.css
@@ -1,5 +1,4 @@
 .songRow {
-  margin-left: 1.5rem;
   padding: 1rem;
   display: flex;
   align-items: center;
@@ -12,11 +11,7 @@
 }
 
 .songRow__info {
-  margin-left: 1.5rem;
-}
-
-.songRow__info > h1 {
-  font-size: 1.2rem;
+  margin-left: 1rem;
 }
 
 .songRow__info > p {
