+++ src/layout/Footer/index.jsx
@@ -35,17 +35,9 @@ function Footer(props) {
       </div>
 
       <div className="footer__right">
-        <Grid container spacing={2}>
-          <Grid item>
-            <PlaylistPlayIcon />
-          </Grid>
-          <Grid item>
-            <VolumeDownIcon />
-          </Grid>
-          <Grid item xs>
-            <Slider />
-          </Grid>
-        </Grid>
+        <PlaylistPlayIcon />
+        <VolumeDownIcon />
+        <Slider />
       </div>
     </div>
   );
