+++ src/layout/Sidebar/SidebarPlaylistsList/SidebarPlaylistsList.css
@@ -1,0 +1,3 @@
+.sidebarPlaylistsList {
+  overflow-y: auto;
+}
